requirements.txt
cv==1.0.0
cycler==0.11.0
docker==4.1.0
docopt==0.6.2
hestia==0.6.0
idna==2.5
joblib==1.2.0
MarkupSafe==2.0.1
matplotlib==3.6.2
marshmallow==3.19.0
nibabel==4.0.2
numpy==1.24.0
opencv-python==4.6.0.66
pathlib==1.0.1
pillow==9.0.1
pytz==2022.7.1
rhea==0.5.5
six==1.13.0
Unipath==1.1
scikit-learn==1.2.0
scikit_image
pytorch_lightning
scipy==1.9.1
monai==1.1.0
medpy
tensorboard
simpleitk
pympler
xvfbwrapper
nibabel
glob2